Working on implementing Window.
HAd a window working, but then realized how it wanted the message handling
method to be static which will cause problems. In the middle of working
that out. Chris used a method of having a static wrapper function that
returned the result of a internal call using the static instance of the
class. I might mimic his code and make my Window Manager a static
instance. Will think on if this is a good idea.