Initialized Graphics.
I don't have a triangle drawing, but I have a basic game system set up.
The basics of it are that Game inherits from GraphicsCore. An instance of
game is created in the Window Manager, which will call the various
functions that are necessary. Next is to render a triangle, by sending a
rendering component to the GraphicsCore. This will set it up so that when
you add objects they are sent to the rendering system, so Game Logic can
continue in the background.