Added the start of the collision system.
Basic design is that you have the collision system have a list of
colliders, collision functions, update function, and a compute mesh
function for each collider mesh type.

Colliders will have a collider mesh type, a collider type, a onCollision
function, a reference to the game objects transform, and the collider
mesh.