Finished basic collision management.
Made it, so that you colliders are generated on mesh loading and then
stored in the asset manager, unaltered. If a game object wants one, they
ask for a copy.

Collision checks account for moving objects. Need to account for scaled
and rotated objects, but the framework is in place.